code/src/requirements.txt
pymongo
python-dotenv
pydantic